Add code gen v1 (#8)
* Prototype conversion to SSA format

* Change storage of operations from vector to map

* Add a couple of missing const annotations

* Add MultiplicationBuilder and clean it up mul logic, add v1 of substitution operation + some tests

* Add support for visitors that accept the input expression as well

* Add AdditionParts, cleanup substitution logic and add more tests

* Add the IsLeaf property to identify leaf expressions vs nested

* Example of simplifying by re-ordering adds+muls (simple count of pairs)

* Create IrBuilder, add some basic tests

* Add Eigen to dependencies

* Add example of code generation test (very rough, needs cleanup)

* Get rid of inheritance of MatrixExpr from Expr

* Fix issue where generate_cpp_test runs every invocation of make

* Move visitors for derivative+diff into cc file

* Improve numeric evaluation, add a couple of simple test examples

* Wrap AST types, rename wrapper directory, add pyproject and test packaging a bit

* Changes to build on Windows

* Wip on refactoring the code generation a bit

* Add basic numeric evaluation, add python code-gen test

+++ test/integer_utils_test.cc
@@ -80,8 +80,8 @@ TEST(IntegerUtils, TestComputePrimeFactors2) {
 
   // Factor numbers 1 -> 100000
   for (const auto& expected_factors : primes) {
-    const int64_t product =
-        std::accumulate(expected_factors.begin(), expected_factors.end(), 1l, std::multiplies<>());
+    const int64_t product = std::accumulate(expected_factors.begin(), expected_factors.end(),
+                                            static_cast<int64_t>(1), std::multiplies<>());
     const std::vector<PrimeFactor> result = ComputePrimeFactors(product);
     ASSERT_TRUE(std::is_sorted(result.begin(), result.end(), PrimeFactorsOrder{}));
     // Check we got the same factors out:
